Deflake test_dataset.py: split torch tests (#32487)
One of the flakiness of test_dataset.py is due to the timeout. This splits out the torch tests from this big test file.

https://github.com/ray-project/ray/issues/32067
